refactor(auth): consolidate OTP verification flow
