+++ src/configs/cookies.ts
@@ -13,9 +13,11 @@ export const COOKIE_KEYS = {
   SANDBOX_INSPECT_ROOT_PATH: 'e2b-sandbox-inspect-root-path',
 } as const
 
+export const COOKIE_MAX_AGE_SECONDS = 60 * 60 * 24 * 365 // 1 year
+
 const BASE_COOKIE_OPTIONS: Partial<ResponseCookie> = {
   path: '/',
-  maxAge: 60 * 60 * 24 * 365, // 1 year
+  maxAge: COOKIE_MAX_AGE_SECONDS,
   sameSite: 'lax',
   secure: process.env.NODE_ENV === 'production',
 }
