+++ tests/unit/proxy-plan.test.ts
@@ -7,12 +7,9 @@ describe('classifyProxyRequest', () => {
     ['/sign-in', 'page-auth', true, true, true],
     ['/sign-up', 'page-auth', true, true, true],
     ['/api/trpc/user.update', 'api-trpc', false, false, false],
-    ['/api/teams/team-slug/metrics', 'api-public', false, false, false],
     ['/api/auth/oauth/session', 'authjs-endpoint', false, false, false],
     ['/api/auth/oauth-start', 'authjs-endpoint', false, false, false],
     ['/api/health', 'api-public', false, false, false],
-    ['/api/team/state', 'api-public', false, false, false],
-    ['/api/sidebar/state', 'api-public', false, false, false],
     ['/docs/quickstart', 'rewrite', false, false, false],
     ['/', 'rewrite', false, false, false],
     ['/unknown-public-page', 'public', false, false, false],
